Move punchout request to server
